Pass width and height as props
The width and height properties may or may not be important to the
execution of a given drawing function. (perhaps only the height is
important etc.)

By passing the props as an object, the implementation of a drawing
method can pick and choose the props it needs, and can also take
advantage of javascript's object destructuring to make the final
implementation more readable as well.