Fix error on draw cleanup
The types for the `cleanUp` method were not set before, and the
conditional that would call `cleanUp` was only checking for null. Since
the cleanUp method could also be void, this was causing void to be
treated as a function, which was firing errors.

The types for cleanUp are more explicit now, and the error should now be
avoidable.